Tidy lp adapter, move to new module
from __future__ import annotations

Restore some deleted test behaviour

Split tests

Revert tests
