+++ src/muse/constraints.py
@@ -94,7 +94,6 @@ technologies:
 from __future__ import annotations
 
 from collections.abc import Mapping, MutableMapping, Sequence
-from dataclasses import dataclass
 from enum import Enum, auto
 from typing import (
     Any,
@@ -110,7 +109,7 @@ import xarray as xr
 from mypy_extensions import KwArg
 
 from muse.registration import registrator
-from muse.timeslices import broadcast_timeslice, distribute_timeslice, drop_timeslice
+from muse.timeslices import broadcast_timeslice, distribute_timeslice
 
 CAPACITY_DIMS = "asset", "replacement", "region"
 """Default dimensions for capacity decision variables."""
@@ -752,479 +751,3 @@ def minimum_service(
         dict(capacity=-capacity, production=production, b=b),
         attrs=dict(kind=ConstraintKind.LOWER_BOUND),
     )
-
-
-def lp_costs(
-    capacity_costs: xr.DataArray,
-    commodities: list[str],
-    timeslice_level: str | None = None,
-) -> xr.Dataset:
-    """Creates dataset of costs for solving with scipy's LP solver.
-
-    Importantly, this also defines the decision variables in the linear program.
-
-    The costs applied to the capacity decision variables are provided. This should
-    have dimensions "asset" and "replacement". In other words, capacity addition
-    is solved for each replacement technology for each existing asset.
-
-    No cost is applied to the production decision variables. Thus, the production
-    component of the resulting dataset is zero, with dimensions determining the
-    production decision variables. This will have dimensions "asset", "replacement",
-    "commodity", and "timeslice". In other words, production is solved for each
-    replacement technology for each existing asset, for each commodity, and for each
-    timeslice.
-
-    Args:
-        capacity_costs: DataArray with dimensions "asset" and "replacement" defining the
-            costs of adding capacity to the system.
-        commodities: List of commodities to create production decision variables for.
-        timeslice_level: The timeslice level of the linear problem.
-    """
-    assert set(capacity_costs.dims) == {"asset", "replacement"}
-
-    # Start with capacity costs as template (defines "asset" and "replacement" dims)
-    production_costs = xr.zeros_like(capacity_costs)
-
-    # Add a "timeslice" dimension, convert multiindex to single index
-    production_costs = broadcast_timeslice(production_costs, level=timeslice_level)
-    production_costs = drop_timeslice(production_costs)
-    production_costs["timeslice"] = pd.Index(
-        production_costs.get_index("timeslice"), tupleize_cols=False
-    )
-
-    # Add a "commodity" dimension
-    production_costs = production_costs.expand_dims(commodity=commodities)
-    assert set(production_costs.dims) == {
-        "asset",
-        "replacement",
-        "commodity",
-        "timeslice",
-    }
-
-    # Result is dataset of provided capacity costs and zero production costs
-    return xr.Dataset(dict(capacity=capacity_costs, production=production_costs))
-
-
-def lp_constraint(constraint: Constraint, lpcosts: xr.Dataset) -> Constraint:
-    """Transforms the constraint to LP data.
-
-    The goal is to create from ``lpcosts.capacity``, ``constraint.capacity``, and
-    ``constraint.b`` a 2d-matrix ``constraint`` vs ``decision variables``.
-
-    #. The dimensions of ``constraint.b`` are the constraint dimensions. They are
-        renamed ``"c(xxx)"``.
-    #. The dimensions of ``lpcosts`` are the decision-variable dimensions. They are
-        renamed ``"d(xxx)"``.
-    #. ``set(b.dims).intersection(lpcosts.xxx.dims)`` are diagonal
-        in constraint dimensions and decision variables dimension, with ``xxx`` the
-        capacity or the production
-    #. ``set(constraint.xxx.dims) - set(lpcosts.xxx.dims) - set(b.dims)`` are reduced by
-        summation, with ``xxx`` the capacity or the production
-    #. ``set(lpcosts.xxx.dims) - set(constraint.xxx.dims) - set(b.dims)`` are added for
-        expansion, with ``xxx`` the capacity or the production
-
-    See :py:func:`muse.constraints.lp_constraint_matrix` for a more detailed explanation
-    of the transformations applied here.
-    """
-    constraint = constraint.copy(deep=False)
-
-    # Deal with timeslice multiindex
-    if "timeslice" in constraint.dims:
-        constraint = drop_timeslice(constraint)
-        constraint["timeslice"] = pd.Index(
-            constraint.get_index("timeslice"), tupleize_cols=False
-        )
-
-    # Rename dimensions in b
-    b = constraint.b.drop_vars(set(constraint.b.coords) - set(constraint.b.dims))
-    b = b.rename({k: f"c({k})" for k in b.dims})
-
-    # Create capacity constraint matrix
-    capacity = lp_constraint_matrix(constraint.b, constraint.capacity, lpcosts.capacity)
-    capacity = capacity.drop_vars(set(capacity.coords) - set(capacity.dims))
-
-    # Create production constraint matrix
-    production = lp_constraint_matrix(
-        constraint.b, constraint.production, lpcosts.production
-    )
-    production = production.drop_vars(set(production.coords) - set(production.dims))
-
-    # Combine data
-    result = xr.Dataset(
-        {"b": b, "capacity": capacity, "production": production}, attrs=constraint.attrs
-    )
-    return result
-
-
-def lp_constraint_matrix(
-    b: xr.DataArray, constraint: xr.DataArray, lpcosts: xr.DataArray
-):
-    """Transforms one constraint block into an lp matrix.
-
-    The goal is to create from ``lpcosts``, ``constraint``, and ``b`` a 2d-matrix of
-    constraints vs decision variables.
-
-     #. The dimensions of ``b`` are the constraint dimensions. They are renamed
-         ``"c(xxx)"``.
-     #. The dimensions of ``lpcosts`` are the decision-variable dimensions. They are
-         renamed ``"d(xxx)"``.
-     #. ``set(b.dims).intersection(lpcosts.dims)`` are diagonal
-         in constraint dimensions and decision variables dimension
-     #. ``set(constraint.dims) - set(lpcosts.dims) - set(b.dims)`` are reduced by
-         summation
-     #. ``set(lpcosts.dims) - set(constraint.dims) - set(b.dims)`` are added for
-         expansion
-     #. ``set(b.dims) - set(constraint.dims) - set(lpcosts.dims)`` are added for
-         expansion. Such dimensions only make sense if they consist of one point.
-
-     The result is the constraint matrix, expanded, reduced and diagonalized for the
-     conditions above.
-
-    """
-    from functools import reduce
-
-    from numpy import eye
-
-    # Sum over all dimensions that are not in the constraint or the decision variables
-    result = constraint.sum(set(constraint.dims) - set(lpcosts.dims) - set(b.dims))
-
-    # Rename dimensions for decision variables
-    result = result.rename(
-        {k: f"d({k})" for k in set(result.dims).intersection(lpcosts.dims)}
-    )
-
-    # Rename dimensions for constraints
-    result = result.rename(
-        {k: f"c({k})" for k in set(result.dims).intersection(b.dims)}
-    )
-
-    # Expand dimensions that are in the decision variables but not in the constraint
-    expand = set(lpcosts.dims) - set(constraint.dims) - set(b.dims)
-    result = result.expand_dims(
-        {f"d({k})": lpcosts[k].rename({k: f"d({k})"}).set_index() for k in expand}
-    )
-
-    # Expand dimensions that are in the constraint but not in the decision variables
-    expand = set(b.dims) - set(constraint.dims) - set(lpcosts.dims)
-    result = result.expand_dims(
-        {f"c({k})": b[k].rename({k: f"c({k})"}).set_index() for k in expand}
-    )
-
-    # Dimensions that are in both the decision variables and the constraint
-    diag_dims = set(b.dims).intersection(lpcosts.dims)
-    diag_dims = sorted(diag_dims)
-
-    if diag_dims:
-
-        def get_dimension(dim):
-            if dim in b.dims:
-                return b[dim].values
-            if dim in lpcosts.dims:
-                return lpcosts[dim].values
-            return constraint[dim].values
-
-        diagonal_submats = [
-            xr.DataArray(
-                eye(len(b[k])),
-                coords={f"c({k})": get_dimension(k), f"d({k})": get_dimension(k)},
-                dims=(f"c({k})", f"d({k})"),
-            )
-            for k in diag_dims
-        ]
-        reduced = reduce(xr.DataArray.__mul__, diagonal_submats)
-        if "d(timeslice)" in reduced.dims:
-            reduced = reduced.drop_vars("d(timeslice)")
-        result = result * reduced
-
-    return result
-
-
-@dataclass
-class ScipyAdapter:
-    """Creates the input for the scipy solvers.
-
-    This adapter is required to convert data (costs and constraints) from a series of
-    dataarrays to a format that can be used by scipy's linear programming solver.
-
-    The scipy solver requires the following inputs as a set of 1D or 2D numpy arrays:
-        c: Coefficients of the linear objective function to be minimized. This is a 1D
-            vector, each element representing the cost of a decision variable.
-        A_ub: The inequality constraint matrix. This is a 2D matrix containing
-            coefficients of the linear inequality constraints, with constraints as rows
-            and decision variables as columns.
-        b_ub: The inequality constraint vector. This is a 1D vector, each element
-            representing an upper bound on the corresponding constraint in A_ub.
-        A_eq: The equality constraint matrix. In practice, since all of the constraints
-            currently implemented are inequality constraints, this will always be zeros.
-            However, this may be changed in the future.
-        b_eq: The equality constraint vector. As above, this will currently always be
-            zeros.
-
-    See the documentation for `scipy.optimize.linprog` for more details:
-    https://docs.scipy.org/doc/scipy/reference/generated/scipy.optimize.linprog.html
-
-    This class takes in costs and constraints as xarray dataarrays, and performs the
-    necessary conversions to create the above inputs. These are saved together in the
-    `kwargs` property, which is passed to the scipy solver. This has an additional
-    attribute `bounds`, which is a tuple of the lower and upper bounds for the
-    decision variables. In practice, we will always use (0, np.inf) as the bounds.
-
-    The output of the scipy solver is a 1D array of decision variables, which must then
-    be converted back into a format that MUSE can understand. To aid this, we
-    save templates for the capacity and production decision variables, which are used
-    to convert the output back into a labelled xarray dataset using the helper function
-    `to_muse`.
-    """
-
-    c: np.ndarray
-    to_muse: Callable[[np.ndarray], xr.Dataset]
-    bounds: tuple[float | None, float | None] = (0, np.inf)
-    A_ub: np.ndarray | None = None
-    b_ub: np.ndarray | None = None
-    A_eq: np.ndarray | None = None
-    b_eq: np.ndarray | None = None
-
-    @classmethod
-    def factory(
-        cls,
-        costs: xr.DataArray,
-        constraints: list[Constraint],
-        commodities: list[str],
-        timeslice_level: str | None = None,
-    ) -> ScipyAdapter:
-        # Calculate costs for the linear problem
-        lpcosts = lp_costs(
-            capacity_costs=costs,
-            commodities=commodities,
-            timeslice_level=timeslice_level,
-        )
-
-        # Create dataset from costs and constraints
-        data = cls._unified_dataset(lpcosts, *constraints)
-
-        # Get capacity constraint matrix / costs
-        capacities = cls._selected_quantity(data, "capacity")
-
-        # Get production constraint matrix / costs
-        productions = cls._selected_quantity(data, "production")
-
-        # Get constraint vector
-        bs = cls._selected_quantity(data, "b")
-
-        # Prepare scipy adapter from constraints
-        kwargs = cls._to_scipy_adapter(capacities, productions, bs, *constraints)
-
-        def to_muse(x: np.ndarray) -> xr.Dataset:
-            return ScipyAdapter._back_to_muse(
-                x,
-                capacity_template=capacities.costs,
-                production_template=productions.costs,
-            )
-
-        return ScipyAdapter(to_muse=to_muse, **kwargs)
-
-    @property
-    def kwargs(self):
-        return {
-            "c": self.c,
-            "A_eq": self.A_eq,
-            "b_eq": self.b_eq,
-            "A_ub": self.A_ub,
-            "b_ub": self.b_ub,
-            "bounds": self.bounds,
-        }
-
-    @staticmethod
-    def _unified_dataset(lpcosts: xr.Dataset, *constraints: Constraint) -> xr.Dataset:
-        """Creates single xr.Dataset from costs and constraints."""
-        from xarray import merge
-
-        # Reformat constraints to lp format
-        lp_constraints = [
-            lp_constraint(constraint, lpcosts) for constraint in constraints
-        ]
-
-        # Rename variables in lp constraints
-        lp_constraints = [
-            constraint.rename(
-                b=f"b{i}", capacity=f"capacity{i}", production=f"production{i}"
-            )
-            for i, constraint in enumerate(lp_constraints)
-        ]
-
-        # Rename dimensions in lpcosts
-        lpcosts = lpcosts.rename({k: f"d({k})" for k in lpcosts.dims})
-
-        # Merge data
-        data = merge([lpcosts, *lp_constraints])
-
-        # An adjustment is required for lower bound constraints
-        for i, constraint in enumerate(constraints):
-            if constraint.kind == ConstraintKind.LOWER_BOUND:
-                data[f"b{i}"] = -data[f"b{i}"]
-                data[f"capacity{i}"] = -data[f"capacity{i}"]
-                data[f"production{i}"] = -data[f"production{i}"]
-
-        # Enusure consistent ordering of dimensions
-        return data.transpose(*data.dims)
-
-    @staticmethod
-    def _selected_quantity(data: xr.Dataset, name: str) -> xr.Dataset:
-        # Select data for the specified quantity ("capacity", "production", or "b")
-        result = cast(
-            xr.Dataset, data[[u for u in data.data_vars if str(u).startswith(name)]]
-        )
-
-        # Rename variables ("costs" for the costs variable, 0/1/2 etc. for constraints)
-        return result.rename(
-            {
-                k: ("costs" if k == name else int(str(k).replace(name, "")))
-                for k in result.data_vars
-            }
-        )
-
-    @staticmethod
-    def _to_scipy_adapter(
-        capacities: xr.Dataset, productions: xr.Dataset, bs: xr.Dataset, *constraints
-    ):
-        """Converts constraints to scipy format.
-
-        The constraints are converted to a format that can be used by scipy's linear
-        programming solver. The constraints are converted to a 2D matrix of constraints
-        vs decision variables. The constraints are then converted to a dictionary that
-        can be used by scipy's linear programming solver.
-
-        Args:
-            capacities: Dataset with decision variables for capacity constraints.
-            productions: Dataset with decision variables for production constraints.
-            bs: Dataset with constraints.
-            *constraints: List of constraints.
-
-        Returns:
-            Dictionary with constraints in a format that can be used by scipy's linear
-            programming solver.
-        """
-
-        def reshape(matrix: xr.DataArray) -> np.ndarray:
-            """Convert constraints matrix to a 2D np array.
-
-            The rows of the constraaints matrix will represent the constraints, and the
-            columns will represent the decision variables.
-            """
-            # Before building LP we need to sort dimensions for consistency
-            if list(matrix.dims) != sorted(matrix.dims):
-                matrix = matrix.transpose(*sorted(matrix.dims))
-
-            # Size of the first dimension
-            # This dimension represents the number of constraints
-            size = np.prod(
-                [matrix[u].shape[0] for u in matrix.dims if str(u).startswith("c")]
-            )
-
-            # Reshape into a 2D array: N constraints x N decision variables
-            return matrix.values.reshape((size, -1))
-
-        def extract_bA(constraints, *kinds: ConstraintKind):
-            """Extracts A and b for constraints of specified kinds.
-
-            These will end up as A_ub and b_ub for inequality constraints, and
-            A_eq and b_eq for equality constraints (see ScipyAdapter).
-            """
-            # Get indices of constraints of the specified kind
-            indices = [i for i in range(len(bs)) if constraints[i].kind in kinds]
-
-            # Convert constraints matrices to 2d np arrays
-            capa_constraints = [reshape(capacities[i]) for i in indices]
-            prod_constraints = [reshape(productions[i]) for i in indices]
-
-            # Convert constraints vectors to 1d
-            constraints_vectors = [
-                bs[i].stack(constraint=sorted(bs[i].dims)) for i in indices
-            ]
-
-            # Concatenate constraints
-            if capa_constraints:
-                A = np.concatenate(
-                    (
-                        np.concatenate(capa_constraints, axis=0),
-                        np.concatenate(prod_constraints, axis=0),
-                    ),
-                    axis=1,
-                )
-                b = np.concatenate(constraints_vectors, axis=0)
-            else:
-                # If there are no constraints of the given kind, return None
-                A = None
-                b = None
-            return A, b
-
-        # Create costs vector by concatenating capacity and production costs
-        c = np.concatenate(
-            (
-                capacities["costs"].values.flatten(),
-                productions["costs"].values.flatten(),
-            ),
-            axis=0,
-        )
-
-        # Extract A and b for inequality constraints
-        A_ub, b_ub = extract_bA(
-            constraints, ConstraintKind.UPPER_BOUND, ConstraintKind.LOWER_BOUND
-        )
-
-        # Extract A and b for equality constraints
-        A_eq, b_eq = extract_bA(constraints, ConstraintKind.EQUALITY)
-
-        # Prepare scipy adapter
-        return {
-            "c": c,
-            "A_ub": A_ub,
-            "b_ub": b_ub,
-            "A_eq": A_eq,
-            "b_eq": b_eq,
-            "bounds": (0, np.inf),
-        }
-
-    @staticmethod
-    def _back_to_muse_quantity(
-        x: np.ndarray, template: xr.DataArray | xr.Dataset
-    ) -> xr.DataArray:
-        """Convert a vector of decision variables to a DataArray.
-
-        Args:
-            x: 1D vector of decision variables, outputted from the scipy solver.
-            template: Template for the decision variables. This may be for either
-                capacity or production variables.
-        """
-        # First create a multidimensional dataarray based on the template
-        result = xr.DataArray(
-            x.reshape(template.shape), coords=template.coords, dims=template.dims
-        )
-
-        # Then rename the dimensions (e.g. "d(asset)" -> "asset")
-        return result.rename({k: str(k)[2:-1] for k in result.dims})
-
-    @staticmethod
-    def _back_to_muse(
-        x: np.ndarray,
-        capacity_template: xr.DataArray,
-        production_template: xr.DataArray,
-    ) -> xr.Dataset:
-        """Convert the full set of decision variables to a Dataset.
-
-        This must have capacity variables first, followed by production variables.
-
-        Args:
-            x: 1D vector of decision variables, outputted from the scipy solver.
-            capacity_template: Template for the capacity decision variables.
-            production_template: Template for the production decision variables.
-        """
-        n_capa = capacity_template.size  # number of capacity decision variables
-
-        capa = ScipyAdapter._back_to_muse_quantity(
-            x[:n_capa], template=capacity_template
-        )
-        prod = ScipyAdapter._back_to_muse_quantity(
-            x[n_capa:], template=production_template
-        )
-        return xr.Dataset({"capacity": capa, "production": prod})
